feat: refactored poll components & supabase
